+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to HPL2-X / Amnesia will be documented in this file.
 ### Engine (HPL2)
 #### Added
 - Live window resizing and renderer reloading for resolution changes
+- Hardware cursor instead of software
 
 #### Changed
 - Retargeted Visual Studio solution to v143 toolset (VS 2022)
@@ -18,6 +19,9 @@ All notable changes to HPL2-X / Amnesia will be documented in this file.
 #### Added
 - "Press any key" prompt to skip pre-menu splash screens.
 - Live window resizing and renderer reloading for resolution changes
+- Simulation Rate option (60/120/144/240 Hz)
+- Uncap FPS option to render as fast as possible instead of locking to the simulation rate
+- Hardware cursor instead of software
 
 #### Changed